requirements.txt
alembic==1.7.6
attrs==21.4.0
backports.zoneinfo==0.2.1
behave==1.2.6
behave-pytest==0.1.1
click==8.0.4
dateparser==1.1.0
dnspython==2.2.0
email-validator==1.1.3
essential-generators==1.0
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
py==1.11.0
PyHamcrest==2.0.3
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
